reorganize testing directory
